fix(Dataviz): Don't hide bars on overflow
