+++ packages/dataviz/src/components/BarChart/HorizontalBarChart/HorizontalBarChart.component.tsx
@@ -9,7 +9,6 @@ import {
 	SECONDARY_BAR_ANIMATION_PROPS,
 	useBarChart,
 } from '../useBarChart.hook';
-import FixedBarSizeWrapper from './FixedHeightBarWrapper/FixedHeightBarWrapper.component';
 import ColoredBar from '../ColoredBar/ColoredBar.component';
 import TooltipCursor from '../TooltipCursor/TooltipCursor.component';
 import { ChartStyle } from '../../../types';
@@ -23,6 +22,10 @@ export interface HorizontalBarChartProps {
 	width?: number;
 	height?: number;
 }
+
+const BAR_GAP = 5;
+const BAR_HEIGHT = 20;
+const X_AXIS_TICKS_HEIGHT = 40;
 
 function HorizontalBarChart({
 	chartStyle,
@@ -38,65 +41,65 @@ function HorizontalBarChart({
 		onBarClick,
 		getTooltipContent,
 	);
-
+	// If no height is provided, guess it from content (we don't want the bars to stretch to available height)
+	const contentHeight = height || X_AXIS_TICKS_HEIGHT + data.length * (BAR_GAP + BAR_HEIGHT);
 	return (
-		<ResponsiveContainer width={width} height={height}>
-			<FixedBarSizeWrapper data={data}>
-				<BarChart
-					layout="vertical"
-					className={styles['horizontal-bar-chart']}
-					onMouseMove={onMouseMove}
-					onClick={onClick}
-					onMouseOut={onMouseOut}
-				>
-					<CartesianGrid strokeDasharray="3 3" horizontal={false} />
+		<ResponsiveContainer width={width} height={contentHeight}>
+			<BarChart
+				data={data}
+				layout="vertical"
+				className={styles['horizontal-bar-chart']}
+				onMouseMove={onMouseMove}
+				onClick={onClick}
+				onMouseOut={onMouseOut}
+			>
+				<CartesianGrid strokeDasharray="3 3" horizontal={false} />
 
-					<Bar
-						dataKey={getPrimaryBarValue}
-						stackId="1"
-						{...PRIMARY_BAR_ANIMATION_PROPS}
-						shape={
-							<ColoredBar
-								chartStyle={chartStyle}
-								focusedBarIndex={focusedBarIndex}
-								barType="primary"
-							/>
-						}
-					/>
+				<Bar
+					dataKey={getPrimaryBarValue}
+					stackId="1"
+					{...PRIMARY_BAR_ANIMATION_PROPS}
+					shape={
+						<ColoredBar
+							chartStyle={chartStyle}
+							focusedBarIndex={focusedBarIndex}
+							barType="primary"
+						/>
+					}
+				/>
 
-					<Bar
-						dataKey={getSecondaryBarValue}
-						stackId="1"
-						{...SECONDARY_BAR_ANIMATION_PROPS}
-						shape={
-							<ColoredBar
-								chartStyle={chartStyle}
-								focusedBarIndex={focusedBarIndex}
-								barType="secondary"
-							/>
-						}
-					/>
+				<Bar
+					dataKey={getSecondaryBarValue}
+					stackId="1"
+					{...SECONDARY_BAR_ANIMATION_PROPS}
+					shape={
+						<ColoredBar
+							chartStyle={chartStyle}
+							focusedBarIndex={focusedBarIndex}
+							barType="secondary"
+						/>
+					}
+				/>
 
-					<Tooltip
-						allowEscapeViewBox={{ x: false, y: true }}
-						isAnimationActive={false}
-						content={<TooltipContent />}
-						cursor={<TooltipCursor dataFeature={dataFeature} />}
-					/>
+				<Tooltip
+					allowEscapeViewBox={{ x: false, y: true }}
+					isAnimationActive={false}
+					content={<TooltipContent />}
+					cursor={<TooltipCursor dataFeature={dataFeature} />}
+				/>
 
-					<XAxis dataKey="value" type="number" orientation="top" />
+				<XAxis dataKey="value" type="number" orientation="top" />
 
-					<YAxis
-						dataKey="key"
-						tickMargin={5}
-						tickLine={false}
-						tick={<ColoredBar.Label chartStyle={chartStyle} focusedBarIndex={focusedBarIndex} />}
-						type="category"
-						interval={0}
-						mirror
-					/>
-				</BarChart>
-			</FixedBarSizeWrapper>
+				<YAxis
+					dataKey="key"
+					tickMargin={5}
+					tickLine={false}
+					tick={<ColoredBar.Label chartStyle={chartStyle} focusedBarIndex={focusedBarIndex} />}
+					type="category"
+					interval={0}
+					mirror
+				/>
+			</BarChart>
 		</ResponsiveContainer>
 	);
 }
